Extract and move SpecimenToolsWebPartFactory.
Extract and move SpecimensPage, making the "Specimens" tab optional. Provide mechanism for registering study tabs, such as "Specimens," from other modules.
