+++ specimen/src/org/labkey/specimen/SpecimenModule.java
@@ -25,7 +25,8 @@ import org.labkey.api.exp.property.PropertyService;
 import org.labkey.api.module.CodeOnlyModule;
 import org.labkey.api.module.ModuleContext;
 import org.labkey.api.security.roles.RoleManager;
-import org.labkey.api.specimen.SpecimenWebPartFactory;
+import org.labkey.api.specimen.view.SpecimenWebPartFactory;
+import org.labkey.api.specimen.SpecimensPage;
 import org.labkey.api.specimen.importer.DefaultSpecimenImportStrategyFactory;
 import org.labkey.api.specimen.model.AdditiveTypeDomainKind;
 import org.labkey.api.specimen.model.DerivativeTypeDomainKind;
@@ -34,7 +35,9 @@ import org.labkey.api.specimen.model.SpecimenDomainKind;
 import org.labkey.api.specimen.model.SpecimenEventDomainKind;
 import org.labkey.api.specimen.model.SpecimenRequestEventType;
 import org.labkey.api.specimen.model.VialDomainKind;
+import org.labkey.api.specimen.view.SpecimenToolsWebPartFactory;
 import org.labkey.api.study.SpecimenService;
+import org.labkey.api.study.StudyService;
 import org.labkey.api.view.WebPartFactory;
 import org.labkey.specimen.security.roles.SpecimenCoordinatorRole;
 import org.labkey.specimen.security.roles.SpecimenRequesterRole;
@@ -58,7 +61,8 @@ public class SpecimenModule extends CodeOnlyModule
     protected Collection<WebPartFactory> createWebPartFactories()
     {
         return List.of(
-            new SpecimenWebPartFactory()
+            new SpecimenWebPartFactory(),
+            new SpecimenToolsWebPartFactory()
         );
     }
 
@@ -86,6 +90,8 @@ public class SpecimenModule extends CodeOnlyModule
     {
         // add a container listener so we'll know when our container is deleted:
         ContainerManager.addContainerListener(new SpecimenContainerListener());
+
+        StudyService.get().registerStudyTabProvider(tabs ->tabs.add(new SpecimensPage("Specimen Data")));
 
         SpecimenService.get().registerSpecimenImportStrategyFactory(new DefaultSpecimenImportStrategyFactory());
