+++ study/api-src/org/labkey/api/specimen/view/SpecimenWebPart.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.labkey.api.specimen;
+package org.labkey.api.specimen.view;
 
 import org.labkey.api.specimen.settings.SettingsManager;
 import org.labkey.api.study.Study;
