+++ study/api-src/org/labkey/api/study/view/ToolsWebPartFactory.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2011-2017 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.labkey.api.study.view;
+
+import org.jetbrains.annotations.NotNull;
+import org.labkey.api.view.BaseWebPartFactory;
+import org.labkey.api.view.HttpView;
+import org.labkey.api.view.Portal;
+import org.labkey.api.view.ViewContext;
+import org.labkey.api.view.WebPartView;
+
+import java.util.List;
+
+/**
+ * User: brittp
+ * Date: Oct 1, 2011 1:49:24 PM
+ */
+public abstract class ToolsWebPartFactory extends BaseWebPartFactory
+{
+    protected ToolsWebPartFactory(String name, @NotNull String defaultLocation, String... additionalLocations)
+    {
+        super(name, defaultLocation, additionalLocations);
+    }
+
+    protected abstract List<StudyToolsWebPart.Item> getItems(ViewContext portalCtx);
+
+    protected abstract String getTitle();
+
+    @Override
+    public WebPartView getWebPartView(@NotNull ViewContext portalCtx, @NotNull Portal.WebPart webPart)
+    {
+        return new StudyToolsWebPart(getTitle(), webPart.getLocation().equals(HttpView.BODY), getItems(portalCtx));
+    }
+}
