+++ study/src/org/labkey/study/view/specimen/autoReportList.jsp
@@ -28,7 +28,7 @@
 <%@ page import="org.labkey.api.view.template.FrameFactoryClassic" %>
 <%@ page import="org.labkey.study.CohortFilter" %>
 <%@ page import="org.labkey.study.CohortFilterFactory" %>
-<%@ page import="org.labkey.study.controllers.specimen.SpecimenController" %>
+<%@ page import="org.labkey.study.controllers.specimen.ReportConfigurationBean" %>
 <%@ page import="org.labkey.study.model.CohortImpl" %>
 <%@ page import="org.labkey.study.model.ParticipantCategoryImpl" %>
 <%@ page import="org.labkey.study.model.ParticipantGroup" %>
@@ -40,8 +40,8 @@
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
-    JspView<SpecimenController.ReportConfigurationBean> me = (JspView<SpecimenController.ReportConfigurationBean>) HttpView.currentView();
-    SpecimenController.ReportConfigurationBean bean = me.getModelBean();
+    JspView<ReportConfigurationBean> me = (JspView<ReportConfigurationBean>) HttpView.currentView();
+    ReportConfigurationBean bean = me.getModelBean();
     Container container = getContainer();
     User user = getUser();
     boolean showCohorts = StudyManager.getInstance().showCohorts(container, user);
